git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/WDK/branches/wsfrestful@14044 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Model/src/java/org/gusdb/wdk/model/config/ModelConfigDB.java
@@ -30,7 +30,7 @@ public abstract class ModelConfigDB implements ConnectionPoolConfig {
   private short maxActive = 20;
   private short maxIdle = 1;
   private short minIdle = 0;
-  private short maxWait = 50;
+  private long maxWait = 50;
 
   /**
    * display DB connection count periodically in the log. This is used to
@@ -168,7 +168,7 @@ public abstract class ModelConfigDB implements ConnectionPoolConfig {
    * @return the maxWait
    */
   @Override
-  public short getMaxWait() {
+  public long getMaxWait() {
     return maxWait;
   }
 
@@ -176,7 +176,7 @@ public abstract class ModelConfigDB implements ConnectionPoolConfig {
    * @param maxWait
    *          the maxWait to set
    */
-  public void setMaxWait(short maxWait) {
+  public void setMaxWait(long maxWait) {
     this.maxWait = maxWait;
   }
